Tests for post and returnAll methods (for Gardener) completed.

+++ src/main/java/com/szklarnia/service/GardenerService.java
@@ -1,5 +1,6 @@
 package com.szklarnia.service;
 
+import com.szklarnia.exception_handler.ApiRequestException;
 import com.szklarnia.model.Gardener;
 import com.szklarnia.model.Greenhouse;
 import com.szklarnia.repository.GardenerRepository;
@@ -25,8 +26,13 @@ public class GardenerService {
     }
 
     //delete gardener by ID
+    //wyjątek!!!!
     public void deleteGardenerById(int gardenerId) {
-       gardenerRepository.deleteById(gardenerId);
+       if(gardenerRepository.existsById(gardenerId)) {
+           gardenerRepository.deleteById(gardenerId);
+       } else {
+           throw new ApiRequestException("Cannot delete non-existing gardener.");
+       }
     }
 
     //get gardener by ID
@@ -35,12 +41,17 @@ public class GardenerService {
     }
 
     //save an entity/post new entity to DB
-    public Optional<Gardener> postNewGardener(Gardener newGardener) {
+    public Gardener postNewGardener(Gardener newGardener) {
+        //jeśli id nie jest null'em, bo ktoś podał id w Postmanie && jeśli istnieje już na bazie, to zwróć błąd, jeśli istnieje to true, jeśli nie false
         if(newGardener.getGardenerId() != null && gardenerRepository.existsById(newGardener.getGardenerId())) {
-            return Optional.empty();
+            throw new ApiRequestException("Cannot post gardener with existing ID.");
         }
-        return Optional.of(gardenerRepository.save(newGardener));
+        return gardenerRepository.save(newGardener);
     }
+
+//    public Gardener saveNewGardener(Gardener gardener) {
+//        return gardenerRepository.save(gardener);
+//    }
 
     //put, entire entity overwritten, wszystkie pola nadpisane w Postmannie, bo będą nulle
     public Optional<Gardener> completeGardenerEntityUpdated(Integer gardenerId, Gardener updateGardener) {
@@ -59,6 +70,7 @@ public class GardenerService {
 
     //w serwisie sprawdzam, czy istnieje taki ogrodnik i szklarnia po ich ID
     //jeśli tak, to dla obiektu ogrodnika gardener ustawić szklarnię setSzklarnia, który przyjmuje szklarnię greenhouse, i zapisać ogrodnika, żeby je złączyć
+    //OneToOne
     public Optional<Gardener> setGreenhouseForGardener(Integer greenhouseId, Integer gardenerId) {
         if(gardenerRepository.existsById(gardenerId) && greenhouseRepository.existsById(greenhouseId)) {
             Gardener gardener = gardenerRepository.findById(gardenerId).get();
